Moved tree sub-components, updated broken unit tests
