+++ src/components/component-names.ts
@@ -79,6 +79,4 @@ export const TIMEPICKER_NAME: string = 'm-timepicker';
 export const TOUCH_NAME: string = 'm-touch';
 export const TOOLTIP_NAME: string = 'm-tooltip';
 export const TREE_NAME: string = 'm-tree';
-export const TREE_ICON_NAME: string = 'm-tree-icon';
-export const TREE_NODE_NAME: string = 'm-tree-node';
 export const VALIDATION_MESSAGE_NAME: string = 'm-validation-message';
