+++ src/components/tree/tree-icon/tree-icon.html
