+++ src/components/tree/tree-icon/tree-icon.scss
@@ -1,4 +1,4 @@
-@import '../../styles/abstracts/variables';
+@import '../../../styles/abstracts/variables';
 
 .m-tree-icon {
     margin-right: $m-spacing--s;
