+++ src/components/tree/tree-icon/tree-icon.ts
@@ -1,11 +1,11 @@
 import { PluginObject } from 'vue';
 import Component from 'vue-class-component';
 import { Prop } from 'vue-property-decorator';
-import { extractExtension } from '../../utils/file/file';
-import { ModulVue } from '../../utils/vue/vue';
+import { extractExtension } from '../../../utils/file/file';
+import { ModulVue } from '../../../utils/vue/vue';
+import IconFilePlugin from '../../icon-file/icon-file';
+import IconPlugin from '../../icon/icon';
 import { TREE_ICON_NAME } from '../component-names';
-import IconFilePlugin from '../icon-file/icon-file';
-import IconPlugin from '../icon/icon';
 import WithRender from './tree-icon.html?style=./tree-icon.scss';
 
 const FOLDER_OPEN: string = 'm-svg__folder-open';
