+++ src/components/tree/tree-node/tree-node.scss
@@ -1,5 +1,5 @@
-@import '../../styles/abstracts/variables';
-@import '../accordion/accordion-transition';
+@import '../../../styles/abstracts/variables';
+@import '../../accordion/accordion-transition';
 
 $m-tree-node--height: $m-touch-size;
 $m-tree-node--child-indentation: 52px;
